add/clean checks to prevent editor messes
