+++ paper/src/main/java/com/badbones69/crazycrates/listeners/items/PaperInteractListener.java
@@ -1,15 +1,19 @@
 package com.badbones69.crazycrates.listeners.items;
 
 import com.badbones69.crazycrates.CrazyCrates;
+import com.badbones69.crazycrates.api.enums.Messages;
 import com.badbones69.crazycrates.api.events.CrateInteractEvent;
 import com.badbones69.crazycrates.tasks.crates.CrateManager;
 import org.bukkit.Location;
 import org.bukkit.block.Block;
+import org.bukkit.entity.Player;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.EventPriority;
 import org.bukkit.event.Listener;
+import org.bukkit.event.block.Action;
 import org.bukkit.event.block.BlockPlaceEvent;
 import org.bukkit.event.player.PlayerInteractEvent;
+import org.bukkit.inventory.EquipmentSlot;
 import org.bukkit.inventory.ItemStack;
 
 public class PaperInteractListener implements Listener {
@@ -18,16 +22,54 @@ public class PaperInteractListener implements Listener {
 
     private final CrateManager crateManager = this.plugin.getCrateManager();
 
+    @EventHandler
+    public void onEditorClick(final PlayerInteractEvent event) {
+        if (event.getHand() == EquipmentSlot.OFF_HAND) return;
+
+        final Block block = event.getClickedBlock();
+
+        if (block == null || block.getType().isAir()) return;
+
+        final Player player = event.getPlayer();
+
+        final Action action = event.getAction();
+
+        if (!this.crateManager.hasEditorCrate(player)) return;
+
+        if (!player.hasPermission("crazycrates.editor")) {
+            this.crateManager.removeEditorCrate(player);
+
+            Messages.force_editor_exit.sendMessage(player, "{reason}", "Lacking permission crazycrates.editor");
+
+            return;
+        }
+
+        final Location location = block.getLocation();
+
+        switch (action) {
+            case RIGHT_CLICK_BLOCK -> this.crateManager.addCrateByLocation(player, location);
+            case LEFT_CLICK_BLOCK -> this.crateManager.removeCrateByLocation(player, location);
+        }
+
+        event.setCancelled(true);
+    }
+
     @EventHandler(priority = EventPriority.HIGHEST)
     public void onPlayerInteractEvent(PlayerInteractEvent event) {
         final Block block = event.getClickedBlock();
 
         if (block == null || block.isEmpty()) return;
 
+        if (this.crateManager.hasEditorCrate(event.getPlayer())) {
+            event.setCancelled(true);
+
+            return;
+        }
+
         final Location location = block.getLocation();
 
         if (this.crateManager.isCrateLocation(location)) {
-            new CrateInteractEvent(event, location).preventUse().callEvent();
+            new CrateInteractEvent(event, location).callEvent();
         }
     }
